Update Option-Status-Alert
we set a trigger in the Apps Script to run this function as frequent as we want. Ideally you would run twice within trading hours. It makes the necessary changes on your options table in G_Sheets. Immediately after each evaluation you will get an alert email with info on current options status changes if there is any. You will be notified if an option got into exercise zone or vice versa.